requirements.txt
tensorflow==2.6.0
tensorflow-addons==0.15.0
gdown==3.6.4
numpy
pandas==1.3.4
scikit-learn==1.0.1
tensorflow-probability==0.15.0
tensorflow-datasets==4.0.1
black
flake8
tqdm
pyyaml>=6.0.0
schema
loguru
pyreadr==0.4.3
keras==2.6.0
compress-pickle
focal-loss
sktime==0.8.1